feat: Implement connor package
This change introduces Rust implementation of connor package. It almost ignores the current Go implementation of the package and sticks to strong types using enums instead of a map[FilterKey]any.

The change also includes some implementation of core package, because this is what connor depends on, specifically on core::Doc.
core::DocumentMapping is also implemented, but it's not used and not tested yet.